add feature that counts and updates remaining spots
